OlgaVoynova Modify tests
